# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LONGEVITY API",
    "version": "0.1.0",
    "description": "Health-score MVP — backend API"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
...
  },
  "paths": {
...
    "/questionnaire": {
      "post": {
        "operationId": "postQuestionnaire",
        "tags": [
          "Sources"
        ],
        "summary": "Self-reported lifestyle/questionnaire values stored under questionnaire source",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "values"
                ],
                "properties": {
                  "values": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "metric",
                        "value",
                        "unit"
                      ],
                      "properties": {
                        "metric": {
                          "type": "string"
                        },
                        "value": {
                          "type": "number"
                        },
                        "unit": {
                          "type": "string"
                        },
                        "measuredAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inserted metrics"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/lifestyle": {
      "post": {
        "operationId": "postLifestyle",
        "tags": [
          "Sources"
        ],
        "summary": "Alias for /questionnaire",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "values"
                ],
                "properties": {
                  "values": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "metric",
                        "value",
                        "unit"
                      ],
                      "properties": {
                        "metric": {
                          "type": "string"
                        },
                        "value": {
                          "type": "number"
                        },
                        "unit": {
                          "type": "string"
                        },
                        "measuredAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inserted metrics"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/labs": {
      "post": {
        "operationId": "postLabs",
        "tags": [
          "Sources"
        ],
        "summary": "Manual lab values",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "values"
                ],
                "properties": {
                  "values": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "metric",
                        "value",
                        "unit"
                      ],
                      "properties": {
                        "metric": {
                          "type": "string"
                        },
                        "value": {
                          "type": "number"
                        },
                        "unit": {
                          "type": "string"
                        },
                        "measuredAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inserted metrics"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/sources/fhir/upload": {
      "post": {
        "operationId": "uploadFhir",
        "tags": [
          "Sources"
        ],
        "summary": "Import lab values from an HL7 FHIR R4 Bundle or Observation (LOINC-mapped, max 5MB)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inserted count"
          },
          "400": {
            "description": "Invalid FHIR resource"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
...
  }
}